% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franken_Core_git.R
\name{BuildKNNGfrkn}
\alias{BuildKNNGfrkn}
\title{Builds k-nearest neighbor graph}
\usage{
BuildKNNGfrkn(fsom, kn, dist_method, silent = FALSE, tSNE = FALSE)
}
\arguments{
\item{fsom}{flowSOM object, as constructed by the ReadInputfrkn function}

\item{kn}{Number of nearest neighbors used to build the mutual KNN graph}

\item{dist_method}{Distance or Similarity function used to build KNN graph. Extended Jaccard Similarity is the recommended default}

\item{silent}{If \code{TRUE}, no progress updates will be printed}

\item{tSNE}{If \code{TRUE}, an alternative tSNE layout is computed as well}
}
\value{
Subsetted and updated flowSOM object
}
\description{
Builds k-nearest neighbor graph
}
