% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franken.R
\name{BuildSOMfrkn}
\alias{BuildSOMfrkn}
\title{Build self-organizing map (borrowed from FlowSOM)}
\usage{
BuildSOMfrkn(fsom, colsToUse = NULL, silent = FALSE, ...)
}
\arguments{
\item{fsom}{}

\item{colsToUse}{}
}
\description{
Build self-organizing map (borrowed from FlowSOM)
}
