% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franken_Core_git.R
\name{colRank}
\alias{colRank}
\title{Rank of a column of a data frame}
\usage{
colRank(df)
}
\arguments{
\item{df}{data frame}
}
\value{
maximum value of a data frame's row
}
\description{
Returns the maximum value of a data frame's row
}
