% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franken_Core_git.R
\name{Franken}
\alias{Franken}
\title{Clusters high dimensional data}
\usage{
Franken(
  readin_files,
  resultsDir,
  marker_ids,
  fseed = NULL,
  xd = 20,
  yd = 20,
  kn = 6,
  dist_method = "ejaccard"
)
}
\arguments{
\item{readin_files}{flowSOM object containing data matrix, generated by running ReadInputfrkn first}

\item{resultsDir}{Directory to store outputs}

\item{marker_ids}{An array of ids of which data columns to use for clustering}

\item{fseed}{Seed to be used for SOM and KNN building step. Setting your own seed is highly encouraged
for reproducibility}

\item{xd}{Width of the SOM grid}

\item{yd}{Height of the SOM grid}

\item{kn}{Number of nearest neighbors used to build the mutual KNN graph.}

\item{dist_method}{Distance or Similarity function used to build KNN graph. Extended Jaccard Similarity 
is the recommended default.}
}
\value{
A \code{list} containing three items: 1) a flowSOM object
        2) metaclustering results on SOM nodes and 3) a matrix of cellsXmarkers outputs

A \code{list} with two items: the first is the flowSOM object 
        containing all information (see the vignette for more detailed 
        information about this object), the second is the metaclustering of 
        the nodes of the grid. This is a wrapper function for 
        \code{\link{ReadInput}}, \code{\link{BuildSOM}}, 
        \code{\link{BuildKNNGfrkn}} and \code{\link{walk_metaclustering}}. 
        Executing them separately may provide more options.
}
\description{
This function takes in single-cell data from fcs files, flowframes, flowsets or data.table/frames
}
