% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franken.R
\name{Franken_data_prep}
\alias{Franken_data_prep}
\title{Prepare data for Franken run}
\usage{
Franken_data_prep(dataDir, file_trim_regex = NULL, file_trim_replace = NULL)
}
\arguments{
\item{dataDir}{Directory where data .fcs files can be found}

\item{file_trim_regex}{Regular expression to identify in the data files' names to be replaced}

\item{file_trim_replace}{Expression to replace in data files' names}
}
\value{
Normalized data (arcsinh factor 5) and modified file names according to user's choice
}
\description{
Arcsinh's data and changes .fcs file names to more useful ones as per the user's choosing
}
