% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrankenTest.R
\name{FrankenTest}
\alias{FrankenTest}
\title{Franken Package Test}
\description{
Contains functions required for Franken clustering procedure
}
\author{
Laura De Vargas Roditi \email{laura.devargasroditi@uzh.ch}
}
