% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franken.R
\name{GetClusters}
\alias{GetClusters}
\title{Get cluster label for every individual cells}
\usage{
GetClusters(fsom)
}
\arguments{
\item{fsom}{FlowSOM object as generated by the Franken function}
}
\value{
vector label for every cell
}
\description{
Get cluster label for every individual cells
}
\examples{
Define path to directory with your .fcs files
DataDir = ''
fileName     = list.files(path = DataDir, pattern = "\\\\.fcs$")
readin_files = ReadInput(fileNames,  scale = FALSE)
frk_ls       = Franken(readin_files, marker_ids=c(9,12,14:18))
cluster_labels = GetClusters(frk_ls)

}
