% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franken.R
\name{rowMax}
\alias{rowMax}
\title{Row-wise max of a data frame}
\usage{
rowMax(df)
}
\arguments{
\item{df}{data frame}
}
\value{
maximum value of a data frame's row
}
\description{
Returns the maximum value of a data frame's row
}
\author{
Laura De Vargas Roditi \email{laura.devargasroditi@uzh.ch}
}
