% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franken.R
\name{update_ReadInput_file}
\alias{update_ReadInput_file}
\title{Updates flowSOM object after data subsetting}
\usage{
update_ReadInput_file(readin, new_ids)
}
\arguments{
\item{readin}{flowSOM object, as constructed by the ReadInputfrkn function}

\item{new_ids}{Row ids of cells to include in subsetted data}
}
\value{
Subsetted and updated flowSOM object
}
\description{
Updates flowSOM object after data subsetting
}
