% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franken.R
\name{walk_metaclustering}
\alias{walk_metaclustering}
\title{Performs walktrap clustering of knn graph on SOM nodes}
\usage{
walk_metaclustering(fsom, saveDir, fseed)
}
\arguments{
\item{fsom}{flowSOM object, as constructed by the ReadInputfrkn function. Make sure to run BuildKNNGfrkn first}

\item{saveDir}{Path to directory in which to save results}

\item{fseed}{Walktrap finds communities in a graph via random walks which means results will differ slightly with each run. With fseed you can set your own seed for reproducilibity}
}
\value{
A file called metaclusterIds.txt containing clustering ids for each cell and a  object including clustering ids for each SOM node
}
\description{
Performs walktrap clustering of knn graph on SOM nodes
}
